requirements.txt
fastapi
uvicorn
pydantic
python-jose
pytest
slowapi
zxcvbn-python
google-auth[requests]
google-auth-oauthlib

